Add terminology support
Terminology (https://github.com/billiob/terminology) is a modern terminal emulator which provides a lot of features. Unfortunately, this terminal does not support the "\033[14t" sequence in order to get the terminal window size in pixel and w3m-img does not seems to work. However, terminology understand the sequence "\033}qs\000" (a "\033[14t" alternative) that returns "W;H;FW;FH\n" where:
* W is the width of the terminal grid in characters
* H is the height of the terminal grid in characters
* FW is the width of 1 character cell in pixels
* FH is the height of 1 character cell in pixels
Using these values, it is thus possible to compute the window height and width in pixels. Moreover, terminology embedded a tool: "tycat" which enable you to display an image in your terminal. By using both, the window size query alternative and tycat as new image_backend, neofetch is now able to deal with terminology.