OpenBSD's battery status get back into the BSD case
Maybe the original commiter forgot the * so s/he decided to get OpenBSD's
case nest into the `$os` instead of `$distro`.

I suspect the original reason that because OpenBSD's `$distro` output
is `OpenBSD $(uname -m)` instead of just `OpenBSD` so the results didn't
appear.

Anyway, for consistency, I decided to bring OpenBSD's battery case back
to the `$distro` case, add a very little fix (which is the *) and the
problem should be fixed.