Ubuntu + Flavors: Use more serious colors.
Don't use defualt colors-we have up to 256 values to test. Make
use of them and have fun!

Xubuntu is sadly a bit too dark-17 - 21 will work so it has to be a pick/choose,
so this may vary over time.

If any whitespaces/empty remove/new exact line edits, that is a whitespace remove.
Atom automatically does that so don't complain at me.