Use XSETTINGS to get theme without a DE
GTK is using XSETTINGS as the primary information for theme, icons and
font. It surpasses what's inside the configuration files. DE will
advertise the values they have in their own registries to XSETTINGS.
Without a DE, users can get XSETTINGS with xsettingsd. It is shipped
with `dump_xsettings` to get the current values.

Unrelated, but without a DE, I doubt the content of gsettings matter
much. GTK will not read it by itself. I would remove it. People
running a part of a DE (gnome-control-center maybe?) will get
XSETTINGS.